+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 + [x] Improved window caption buttons
 + [x] Improved `ribbon-menu` component, added scrollable feature
 + [x] Removed default transition property, update loading button
++ [x] Fix bug in `countdown` component
 
 ### 5.1.0
 + [x] Updated color values, each color is worked by hand.